Move regression steps to shared mix-in
